feat: Multithreading
